Complete user + station + role + edit user login adminlayout

+++ BusBooking_Project/Models/Entities/Account.cs
@@ -24,7 +24,7 @@ namespace BusBooking_Project.Models.Entities
         public DateTime? DayEdited { get; set; }
         public int? EditerId { get; set; }
         public bool? Status { get; set; }
-        public byte? RoleId { get; set; }
+        public int? RoleId { get; set; }
         public bool? Active { get; set; }
         public string Description { get; set; }
         public string ForgotPass { get; set; }
